m12n: Move BlobByteSource to oak-lucene
Only oak-lucene consumes that class, no point in moving to oak-commons
